Reorganized Post Token Route in different files and folders
